Remove a redundant cast

+++ clox/src/compiler.c
@@ -93,7 +93,7 @@ error_at(Token *token, const char *message)
   if (token->type == TOKEN_EOF)
     fprintf(stderr, " at end");
   else if (token->type != TOKEN_ERROR)
-    fprintf(stderr, " at '%.*s'", (int) token->length, token->start);
+    fprintf(stderr, " at '%.*s'", token->length, token->start);
   fprintf(stderr, ": %s\n", message);
   parser.had_error = true;
 }
